betterbeauty/requirements/common.txt
appnope==0.1.0
attrs==17.4.0
backcall==0.1.0
boto==2.48.0
boto3==1.7.2
botocore==1.10.2
coverage==4.5.1
decorator==4.2.1
dj-database-url==0.5.0
Django==2.0.3
django-annoying==0.10.4
django-cors-headers==2.2.0
django-custom-user==0.7
django-dynamic-fixture==2.0.0
django-extensions==2.0.6
django-model-utils==3.1.1
django-rest-framework==0.1.0
django-storages==1.6.6
django-timezone-field==2.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
docutils==0.14
facebook-sdk==2.0.0
flake8==3.5.0
flake8-import-order==0.17.1
freezegun==0.3.10
ipython==6.3.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.11.1
jmespath==0.9.3
mccabe==0.6.1
mock==2.0.0
more-itertools==4.1.0
mypy==0.580
parso==0.1.1
path.py==11.0.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.1.0
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
py==1.5.3
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.6.1
PyYAML==3.12
pytest==3.5.0
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-mock==1.10.0
python-dateutil==2.6.1
pytz==2018.3
raven==6.9.0
requests==2.9.1
s3transfer==0.1.13
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
typed-ast==1.1.0
wcwidth==0.1.7
